Inline merkle tree
